Profile now automatically redirects you to login if you try to access the profile page manually while logged out

+++ main.py
@@ -140,50 +140,57 @@ def accountcreatedsuccess():
 
 @app.route('/profile', methods=['GET', 'POST'])
 def profile():
-    # Retrieve the username from the session or set it to None if the key is missing
-    usernameP = session.get('username')
-    loggedIn = session.get('loggedIn')
-    errorText = "placeholder"
 
-    loggedOut = None
-    deleteAccount = None
+    print(session.get('loggedIn'))
 
-    #loggedin detected
+    if session.get('loggedIn') == True:
+        # Retrieve the username from the session or set it to None if the key is missing
+        usernameP = session.get('username')
+        loggedIn = session.get('loggedIn')
+        errorText = "placeholder"
 
+        loggedOut = None
+        deleteAccount = None
 
-    if request.method == 'POST':
-        # Retrieve the value of the 'logOut' form field
-
-        loggedOut = request.form.get('logOut')
-        deleteAccount = request.form.get('deleteAccount')
+        #loggedin detected
 
 
-        noLoginYet = "You haven't logged in!"
-        accountDeleted = "Account successfully deleted..."
+        if request.method == 'POST':
+            # Retrieve the value of the 'logOut' form field
 
-        if loggedOut == "True":
-            # Remove the 'username' key from the session if the user logs out
-            session['username'] = None
-            session['loggedIn'] = False
-            errorText = "Logged out."
-
-            return redirect(url_for("login"))
+            loggedOut = request.form.get('logOut')
+            deleteAccount = request.form.get('deleteAccount')
 
 
-        if deleteAccount == "True": #works
-            deleteQuery = "DELETE FROM mysql.registeredAccounts WHERE user_name = %s"
-            mycursor.execute(deleteQuery, (session['username'],))
-            db.commit()
+            noLoginYet = "You haven't logged in!"
+            accountDeleted = "Account successfully deleted..."
 
-            session['username'] = None
-            session['loggedIn'] = False
-            errorText = "Account deleted successfully."
+            if loggedOut == "True":
+                # Remove the 'username' key from the session if the user logs out
+                session['username'] = None
+                session['loggedIn'] = False
+                errorText = "Logged out."
 
-            # DO NOT DO THIS, THIS DIRECTLY TAMPERS THE CODE, USE HTML/JS
-            # if loggedOut == "True":
-            #     session['username'] = "You can't delete now! Youve already logged out..."
+                return redirect(url_for("login"))
 
-    return render_template('profile.html', usernameP=usernameP, loggedIn=loggedIn, errorText=errorText)
+
+            if deleteAccount == "True": #works
+                deleteQuery = "DELETE FROM mysql.registeredAccounts WHERE user_name = %s"
+                mycursor.execute(deleteQuery, (session['username'],))
+                db.commit()
+
+                session['username'] = None
+                session['loggedIn'] = False
+                errorText = "Account deleted successfully."
+
+                # DO NOT DO THIS, THIS DIRECTLY TAMPERS THE CODE, USE HTML/JS
+                # if loggedOut == "True":
+                #     session['username'] = "You can't delete now! Youve already logged out..."
+
+        return render_template('profile.html', usernameP=usernameP, loggedIn=loggedIn, errorText=errorText)
+
+    else:
+        return redirect(url_for("login"))
 
 
 app.run(debug=True)
